feat: Support customizable pulse frequencies for likes
Introduces the ability to specify and store custom pulse frequencies when a user likes a post. This enhances the expressiveness of user interactions and allows for more nuanced engagement signals within the platform.

The `pulseFrequency` field has been added to the `AppNotification` type and the `handleLike` function has been updated to accept and store this new parameter. The UI also displays the frequency in the notification.